requirements.txt
telethon
redis
python-dotenv
pydantic[email]
pillow
imagehash
asyncpg
aiosmtplib
